requirements.txt
# Core dependencies
langgraph>=0.2.0
langfuse>=2.0.0
ollama>=0.3.0

# MCP (Model Context Protocol) integration
mcp>=1.0.0

# Configuration and environment
python-dotenv>=1.0.0

# Data handling
pydantic>=2.5.0

# Logging and monitoring
structlog>=23.2.0

# Code validation and formatting
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development tools
pre-commit>=3.6.0

# Visualization (optional)
graphviz>=0.20.0
pillow>=10.0.0
matplotlib>=3.7.0